Fix breaking change in auth flow.
Previously, /one-account-user/api/v1/identity-providers expected
an email and auth type, which it no longer expects to see. Additionally,
it expects a new argument `countryCode`. This change fixes the auth flow
by passing the new parameters and adds an optional `country_code`
argument to the Frigidaire constructor, which defaults to "US".